requirements.txt
numpy
colour-science
matplotlib
scikit-image
pillow
opencv-python-headless